Leetcode 1. Two Sum (VS Code)
