+++ .gitignore
@@ -2,6 +2,7 @@
 .DS_Store
 .vscode/
 *.code-workspace
+build/
 
 # XCode
 [Xx]cuserdata/
